refactor main.py
