+++ caesar.py
@@ -1,0 +1,16 @@
+def rot13(letter):
+    if 'A' <= letter <= 'Z':
+        offset = ord('A')
+    elif 'a' <= letter <= 'z':
+        offset = ord('a')
+    else:
+        return letter
+
+    offset_letter = ord(letter) - offset + 13
+    wrapped_letter = offset_letter % 26
+    new_letter = chr(wrapped_letter + offset)
+
+    return new_letter
+
+def caesar(msg):
+    return "".join([rot13(letter) for letter in msg])
