+++ main.py
@@ -1,3 +1,6 @@
+# from caesar import rot13
+from caesar import caesar
+
 def crypt_letter(letter, key_letter):
     return chr(ord(letter) ^ ord(key_letter))
 
@@ -5,22 +8,6 @@ def crypt(msg, key):
     key_len = len(key)
     return "".join([crypt_letter(letter, key[i % key_len]) for i, letter in enumerate(msg)])
 
-def rot13(letter):
-    if 'A' <= letter <= 'Z':
-        offset = ord('A')
-    elif 'a' <= letter <= 'z':
-        offset = ord('a')
-    else:
-        return letter
-
-    offset_letter = ord(letter) - offset + 13
-    wrapped_letter = offset_letter % 26
-    new_letter = chr(wrapped_letter + offset)
-
-    return new_letter
-
-def caesar(msg):
-    return "".join([rot13(letter) for letter in msg])
     
 def main():
     print(f"{caesar('Hello')=}")
